requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cbor==1.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.7
djangorestframework==3.14.0
gensim==4.3.2
idna==3.6
ijson==3.2.3
inscriptis==2.3.2
ir-datasets==0.5.5
joblib==1.3.2
lightgbm==4.1.0
lxml==4.9.3
lz4==4.3.2
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
pyautocorpus==0.1.12
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
trec-car-tools==2.6
typing_extensions==4.8.0
tzdata==2023.3
unlzw3==0.2.2
urllib3==2.1.0
warc3-wet==0.2.3
warc3-wet-clueweb09==0.2.5
zlib-state==0.1.6
